base case

+++ performance.py
@@ -7,8 +7,8 @@ def calculate_sharpe_ratio(returns, periods=252):
 def calculate_drawdowns(equity_curve):
     hwm = [0]
     eq_index = equity_curve.index
-    drawdown = pd.Series(index=eq_index)
-    duration = pd.Series(index=eq_index)
+    drawdown = pd.Series(index=eq_index, dtype=np.float64)
+    duration = pd.Series(index=eq_index, dtype=np.int64)
 
     for i in range(1, len(eq_index)):
         current_hwm = max(hwm[i-1], equity_curve[i])
